+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+## 1.1.0-nullsafety.3
+
+* Added `stringBeforeLength` and `stringAfterLength` to `CharacterRange`.
+* Added `CharacterRange.at` constructor.
+* Added `getRange(start, end)` and `characterAt(pos)` to `Characters`
+  as alternative to `.take(end).skip(start)` and `getRange(pos, pos + 1)`.
+* Change some positional parameter names from `other` to `characters`.
+
 ## 1.1.0-nullsafety.2
 
 * Update for the 2.10 dev sdk.
-
 
 ## 1.1.0-nullsafety.1